refactor : ChessBoardFactory 삭제 및 ChessBoard 정적 팩토리 메서드 추가
