+++ src/main/java/domain/chessboard/ChessBoard.java
@@ -7,10 +7,26 @@ import java.util.List;
 
 public class ChessBoard {
 
+    private static final int NUMBER_OF_NONE_LINES = 4;
+
     private final List<Rank> chessBoard;
 
     public ChessBoard(final List<Rank> chessBoard) {
         this.chessBoard = chessBoard;
+    }
+
+
+    public static ChessBoard generate() {
+        final List<Rank> chessBoard = new ArrayList<>();
+        chessBoard.add(Rank.initRankToRank(InitRank.OTHERS_BLACK));
+        chessBoard.add(Rank.initRankToRank(InitRank.PAWN_BLACK));
+        for (int i = 0; i < NUMBER_OF_NONE_LINES; i++) {
+            chessBoard.add(Rank.initRankToRank(InitRank.EMPTY));
+        }
+        chessBoard.add(Rank.initRankToRank(InitRank.PAWN_WHITE));
+        chessBoard.add(Rank.initRankToRank(InitRank.OTHERS_WHITE));
+
+        return new ChessBoard(new ArrayList<>(chessBoard));
     }
 
     public List<Rank> getChessBoard() {
