+++ src/test/java/domain/ChessBoardTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import domain.chessboard.ChessBoardFactory;
+import domain.chessboard.ChessBoard;
 import domain.chessboard.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,12 +9,12 @@ import java.util.Collection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class ChessBoardFactoryTest {
+public class ChessBoardTest {
 
     @Test
     @DisplayName("체스판을 생성한다.")
     void givenChessBoard_thenSize64() {
-        final int sum = ChessBoardFactory.generate()
+        final int sum = ChessBoard.generate()
                 .getChessBoard()
                 .stream()
                 .map(Rank::getRank)
